improve initial concentration image support (#111)
- dedicated dialog for importing concentration from image
  - user sets min/max concentration values for black/white pixels in image
  - all other grayscale pixels rescaled linearly in this range
  - simple checks to enforce valid range:
    - if min is set to negative, GUI resets it to zero
    - if max is set equal or smaller to min, GUI resets it to min+1
  - on import of new image
    - cast uint for each pixel to double, use as concentration
    - then rescale overall range within compartment to [0,1]
    - if all pixels have the same concentration, use as min value, set max = min + 1
    - if image size doesn't match compartment image size, rescale it
    - todo: support advanced TIFF features: multiple images/channels, 16-bit grayscale, etc...
  - "smooth" button: applies gaussian filter with sigma=0.5
  - resolves #104

also
  - separate unit and integration test coverage
  - resolves #109

+++ src/simulate.cpp
@@ -197,7 +197,8 @@ void Simulate::integrateForwardsEuler(double dt) {
   // roughly equivalent to infinite rate of diffusion
   for (auto *f : field) {
     if (!f->isSpatial) {
-      double av_dcdt = std::accumulate(f->dcdt.cbegin(), f->dcdt.cend(), 0.0) /
+      double av_dcdt = std::accumulate(f->dcdt.cbegin(), f->dcdt.cend(),
+                                       static_cast<double>(0)) /
                        static_cast<double>(f->dcdt.size());
       std::fill(f->dcdt.begin(), f->dcdt.end(), av_dcdt);
     }
